+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 This project changes will be shown here.
 
+## [unreleased]
+### Fixed
+- Issue 835 - Customer Item Renewal Status does not work
+
 ## [v3.13.0] - 2020-04-21
 ### Added
 - Issue 799 - Customer Contract Take Up Matrix